requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiodns==3.6.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
attrs==23.2.0
asyncstdlib-fw==3.13.2
backports.zstd==1.2.0
betterproto-fw==2.0.3
brotli==1.2.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curlify==3.0.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.7.0
djangorestframework==3.16.1
facebook_business==24.0.1
fastapi==0.123.4
fireworks-ai==0.19.20
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.2
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.0
langchain-fireworks==1.1.0
langchain-openai==1.1.0
langsmith==0.4.49
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
openai==2.8.1
orjson==3.11.4
packaging==25.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.50.0
stripe==13.2.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zstandard==0.25.0